+++ fidimag/micro/baryakhtar.py
@@ -1,28 +1,24 @@
 import numpy as np
 import fidimag.extensions.baryakhtar_clib as clib
-from .micro_driver import MicroDriver
+from fidimag.common.llg_driver import LLG_Driver
 from .relax import Relaxation
 from .relax import Laplace
 
 
-class LLBarFull(MicroDriver):
+class LLBarFull(LLG_Driver):
 
-    def __init__(self, mesh, spin, Ms, field, alpha, pins,
-                 interactions,
-                 name,
-                 data_saver,
-                 integrator='sundials',
-                 use_jac=False,
-                 chi=1e-3
-                 ):
-
+    def __init__(self, mesh, spin, magnitude, pins, 
+                interactions, 
+                field, 
+                data_saver, integrator = "sundials", 
+                use_jac=False,
+                chi=1e-3):
         # Inherit from the driver class
-        super(LLBarFull, self).__init__(mesh, spin, Ms, field,
-                                        alpha, pins, interactions, name,
+        super(LLBarFull, self).__init__(mesh, spin, magnitude, pins, 
+                                        interactions, field, 
                                         data_saver,
-                                        integrator='sundials',
-                                        use_jac=False
-                                        )
+                                        integrator = integrator, 
+                                        use_jac=False)
 
         self.chi = chi
         self.lap = Laplace(mesh)
@@ -56,23 +52,21 @@ class LLBarFull(MicroDriver):
         return 0
 
 
-class LLBar(MicroDriver):
+class LLBar(LLG_Driver):
 
-    def __init__(self, mesh, spin, Ms, field, alpha, pins,
-                 interactions,
-                 name,
-                 data_saver,
-                 integrator='sundials',
-                 use_jac=False,
-                 ):
-
+    def __init__(self, mesh, spin, magnitude, pins, 
+                interactions, 
+                field, 
+                data_saver, integrator = "sundials", 
+                use_jac=False,
+                chi=1e-3):
         # Inherit from the driver class
-        super(LLBar, self).__init__(mesh, spin, Ms, field,
-                                    alpha, pins, interactions, name,
-                                    data_saver,
-                                    integrator='sundials',
-                                    use_jac=False
-                                    )
+        super(LLBar, self).__init__(mesh, spin, magnitude, pins, 
+                                        interactions, field, 
+                                        data_saver,
+                                        integrator = integrator, 
+                                        use_jac=False)
+
 
         self.lap = Laplace(mesh)
