+++ fidimag/micro/llg.py
@@ -1,14 +1,16 @@
 from __future__ import division
 from __future__ import print_function
 
+from fidimag.common.llg_driver import LLG_Driver
+
 # We use the atomistic/lib/llg.c file to calculate the LLG equation for
 # the micromagnetic case -> Move this library to Common in the future
 import fidimag.extensions.clib as clib
 
-from .micro_driver import MicroDriver
+#from .micro_driver import Driver
 
 
-class LLG(MicroDriver):
+class LLG(LLG_Driver):
     """
 
     This class is the driver to solve the Landau Lifshitz Gilbert equation
@@ -28,21 +30,21 @@ class LLG(MicroDriver):
 
     """
 
-    def __init__(self, mesh, spin, Ms, field, alpha, pins,
-                 interactions,
-                 name,
-                 data_saver,
-                 integrator='sundials',
-                 use_jac=False
-                 ):
+    def set_default_options(self, gamma=2.21e5, Ms=8.0e5, alpha=0.1):
+        """
+        Default option for the integrator
+        Default gamma is for a free electron
+        """
+        self.default_c = 1e11
+        self._alpha[:] = alpha
 
-        # Inherit from the driver class
-        super(LLG, self).__init__(mesh, spin, Ms, field,
-                                  alpha, pins, interactions, name,
-                                  data_saver,
-                                  integrator='sundials',
-                                  use_jac=False
-                                  )
+        # When we create the simulation, Ms is set to the default value. This
+        # is overriden when calling the set_Ms method from the Siulation class
+        # or when setting Ms directly (property)
+        self._magnitude[:] = Ms
+
+        self.gamma = gamma
+        self.do_precession = True
 
     def sundials_rhs(self, t, y, ydot):
         """
